requirements.txt
pydantic
openai
python-dotenv
flask
pyyaml
gunicorn
